fix: Replace f64 with Decimal for all financial calculations
- Replace f64 with Decimal in token/pools.rs for liquidity_usd
- Update token/metadata.rs liquidity functions to return Decimal
- Replace f64 with Decimal in engine/volume_cache.rs for volume_usd
- Update engine/market_regime.rs to use Decimal for volume calculations
- Update engine/signal_quality.rs to accept Decimal for liquidity
- Update handlers/webhook.rs to pass Decimal for liquidity
- Remove unnecessary f64->Decimal conversions in token/parser.rs

This ensures all financial calculations (PnL, Amount, Price, Liquidity, Volume)
use rust_decimal to avoid IEEE 754 floating-point precision errors.

Fixes floating-point math issue where f64 was used for currency calculations.

+++ operator/src/engine/momentum_exit.rs
@@ -107,7 +107,7 @@ impl MomentumExit {
 
         // Check 2: Volume drop (>50% from 24h average)
         if let Some(ref volume_cache) = self.volume_cache {
-            if volume_cache.has_volume_drop(token_address, 50.0) {
+            if volume_cache.has_volume_drop(token_address, Decimal::from(50)) {
                 tracing::warn!(
                     trade_uuid = %trade_uuid,
                     token_address = token_address,
